+++ ft_calloc.c
@@ -6,16 +6,16 @@
 /*   By: fbes <fbes@student.codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/10/26 21:00:09 by fbes          #+#    #+#                 */
-/*   Updated: 2020/10/26 21:27:08 by fbes          ########   odam.nl         */
+/*   Updated: 2020/10/27 14:24:11 by fbes          ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_calloc(unsigned int nmemb, unsigned int size)
+void	*ft_calloc(size_t nmemb, size_t size)
 {
 	void			*ptr;
-	unsigned int	size_to_malloc;
+	size_t			size_to_malloc;
 
 	if (nmemb == 0 || size == 0)
 		return (malloc(1));
