+++ ft_memset.c
@@ -6,11 +6,13 @@
 /*   By: fbes <fbes@student.codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/10/26 15:29:31 by fbes          #+#    #+#                 */
-/*   Updated: 2020/10/26 16:23:13 by fbes          ########   odam.nl         */
+/*   Updated: 2020/10/27 14:19:38 by fbes          ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *s, int c, unsigned int n)
+#include "libft.h"
+
+void	*ft_memset(void *s, int c, size_t n)
 {
 	char	*copy;
