+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: fbes <fbes@student.codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/10/26 19:24:26 by fbes          #+#    #+#                 */
-/*   Updated: 2020/10/26 19:27:52 by fbes          ########   odam.nl         */
+/*   Updated: 2020/10/27 14:31:22 by fbes          ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t		ft_strlen(const char *s)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
